CLI補完高速化 (import減らしたり遅延させたり)
