--- .env
@@ -1,1 +1,0 @@
-NEO4J_VERSION=5.20
