+++ knowde/complex/systats/__init__.py
@@ -1,0 +1,1 @@
+"""ネットワーク評価."""
