+++ pytest.ini
@@ -16,6 +16,6 @@ norecursedirs =
   .venv
 env =
   NEO4J_USERNAME=neo4j
-  NEO4J_PASSWORD=password
+  NEO4J_PASSWORD=passwordneo
   NEO4J_URI=localhost
   NEO4J_URL=bolt://{NEO4J_USERNAME}:{NEO4J_PASSWORD}@{NEO4J_URI}:7687
